+++ priv/template/lib/application_name.ex
@@ -7,8 +7,11 @@ defmodule <%= application_module %> do
     import Supervisor.Spec, warn: false
 
     children = [
-      # Define workers and child supervisors to be supervised
-      # worker(<%= application_module %>.Worker, [arg1, arg2, arg3])
+      # Start the endpoint when the application starts
+      worker(<%= application_module %>.Endpoint, []),
+
+      # Here you could define other workers and supervisors as children
+      # worker(<%= application_module %>.Worker, [arg1, arg2, arg3]),
     ]
 
     # See http://elixir-lang.org/docs/stable/elixir/Supervisor.html