+++ priv/template/test/test_helper.exs
@@ -1,2 +1,1 @@
-<%= application_module %>.Endpoint.start
 ExUnit.start