+++ test/mix/phoenix_test.exs
@@ -1,12 +1,13 @@
 defmodule Mix.PhoenixTest do
   use ExUnit.Case, async: true
 
-  test "router/0 returns the router based on the Mix application" do
-    assert Mix.Phoenix.router == Phoenix.Router
+  test "base/0 returns the module base based on the Mix application" do
+    assert Mix.Phoenix.base == "Phoenix"
   end
 
-  test "endpoint/0 returns the router based on the Mix application" do
-    assert Mix.Phoenix.endpoint == Phoenix.Endpoint
+  test "endpoints/0 returns the endpoints based on the Mix application" do
+    assert Mix.Phoenix.endpoints([]) == [Phoenix.Endpoint]
+    assert Mix.Phoenix.endpoints(["Hello.Endpoint"]) == [Hello.Endpoint]
   end
 
   test "modules/0 returns all modules in project" do